Trabajo completo

+++ TotalTalentFinalProyect/src/main/java/Modelo/Factory/ContratoParcial.java
@@ -2,18 +2,23 @@ package Modelo.Factory;
 
 import Modelo.Contrato;
 
+// --- Clase ContratoParcial (Producto Concreto) ---
+// Esta es otra subclase de Contrato (herencia).
+// Representa un contrato de tipo "Parcial" o por horas.
+// Es otro "producto" que el ContratoFactory puede crear.
 public class ContratoParcial extends Contrato {
 
-    // Atributos específicos para el contrato parcial con horas trabajadas y su pago por hora
+    // Atributos específicos para el contrato parcial.
     private int horasTrabajadas;
-    private double pagoPorHora;
+    private float pagoPorHora;
 
-    // Constructor para la clase ContratoParcial
+    // Constructor para la clase ContratoParcial.
     public ContratoParcial() {
+        // Llama al método heredado para auto-asignarse su tipo.
         this.setTipoContrato("Parcial");
     }
 
-    // Métodos getter y setter para horasTrabajadas y pagoPorHora
+    // Métodos getter y setter para los atributos específicos.
     public int getHorasTrabajadas() {
         return horasTrabajadas;
     }
@@ -22,18 +27,23 @@ public class ContratoParcial extends Contrato {
         this.horasTrabajadas = horasTrabajadas;
     }
 
-    public double getPagoPorHora() {
+    public float getPagoPorHora() {
         return pagoPorHora;
     }
 
-    public void setPagoPorHora(double pagoPorHora) {
+    public void setPagoPorHora(float pagoPorHora) {
         this.pagoPorHora = pagoPorHora;
     }
 
-    // Método para calcular el salario mensual basado en las horas trabajadas y el pago por hora
+    // --- Implementación del Método Abstracto ---
+    // Sobrescribe el método abstracto calcularSueldo heredado de Contrato.
+    // Implementa la lógica de cálculo específica para un contrato parcial.
     @Override
-    public double calcularSueldo() {
-        return horasTrabajadas * pagoPorHora + getBonificacion(); // No se le quita AFP
+    public float calcularSueldo() {
+        // La fórmula es (horas * pago_por_hora) + bonificación.
+        // Nótese que este tipo de contrato no usa getSalarioBase()
+        // ni resta el descuento de AFP en su cálculo.
+        return horasTrabajadas * pagoPorHora + getBonificacion();
     }
 
 }
